move helper closer to event classes
