+++ plugins/gradle/src/org/jetbrains/plugins/gradle/execution/test/runner/events/ConfigurationErrorEvent.java
@@ -29,7 +29,6 @@ import org.jetbrains.plugins.gradle.GradleManager;
 import org.jetbrains.plugins.gradle.execution.test.runner.GradleTestsExecutionConsole;
 import org.jetbrains.plugins.gradle.service.project.GradleNotification;
 import org.jetbrains.plugins.gradle.util.GradleConstants;
-import org.jetbrains.plugins.gradle.util.XmlXpathHelper;
 
 import javax.swing.event.HyperlinkEvent;
